Prototype testing code
This commit contains the quick and dirty code that was sketched
to test the circuit after first assembly. It is not much,
but it vaguely shows the structure I want to build towards.

The general idea is to have a continuous loop where all light effects
are defined as a general function of the time and location of each LED
which is then continuously evaluated over the whole mirror.