Introduce AbstractEffect
This commit creates the abstract base class for all effects, with
virtual methods to randomize and evaluate the color for each led.

The couple of effects that I had written to test the circuit are also
rewritten to use the new base class. Do they work? Who knows. Valentin
was supposed to bring me the mirror a couple of hours ago, but I guess I
will find out another day.