Rename magic macro variables, comments pass
Cleanup commit to add comments here and there and a few TODOs.
The only code change is that the variables created by the FOR_EACH_
macros are renamed to make them uglier, which also reduces the chance
for an accidental name collision 3 months from now when I have forgotten
everything about it.