Add precompute method to improve performance
The downside of the shader-like approach is that sometimes you end up
having to repeat calculations because the function is applied per pixel.
This was obvious in the looping point demo effect, where the color was
being recomputed every time for each led even though it was the same.

This commit adds a precompute method to the abstract effect class.
It's called at the start of the frame and allow the effect to precompute
shared values for the frame.

In my test with looping point it raised the fps from 90 to 125!