Add postprocess method, improve LoopingPoint
This commit adds a postprocess() method for the effects that is called
after all the leds have been evaluate()ed, to allow for fades and
similar effects.

Demonstrated in the LoopingPoint effect: one led is turned on fully and
the others are continuously faded, creating a trail behind the leading
led.