Add support for animations
This commit introduces a base animation class that can be used to write
short programs. Animations are much more free, in the sense that the
controller calls their run() method and they are free to do whatever
they want without any of that fancy math ideas for effects.

The idea is that they should represent single run of non-repeating
patterns that can be used as transitions between different effects.

Close #2