My constant struggle with random()
I've been having a lot of issues with random being seemingly not random
lately, which explained the stupid analogRead calls.

This commit reverts to the correct invocations of random, which seems to
work fine now for some reason. I have no idea what was going on but
whatever that is it's not happening anymore today.

... unless you try to generate a random HSV color. That always results
in the same color. Investigation pending.

I also added an unconnected jumper sitting next to an electrical
connection for good measure, hoping to pick up more electrical noise.