Improve both the code and the look of ElectricSparks
Simplifies and improves the logic to roll for larger sparks.
There is now a single chance parameter that controls the doubling of the
width. The effect keeps rolling this dice until it fail, doubling the
size each time.

This makes for prettier code but luckily also for a prettier effect. It
somehow feels a bit more organic and produces a wider range of sparks.