More unique logic for the Glow effect
It was pretty lame in person. This commit changes its logic to be more
interesting: instead of a single color pulsating in brightness, it now
pulses around a central hue always at full brightness. It also uses a
cubicwave instead so that it spends more time around the extremes, which
feels more like breathing.