Simplify the logic for forced FX/animations
Instead of keeping a pointer to a preinstatiated effect or animation,
this commit lets you force the selection index instead.

This lets us simplify the overall control logic, since there is no need
to handle separately the forced effect or animation, and since it always
creates a new instance for both, it also closes #17.

While simplifying the control logic I also removed the first animation
and choice of effect from the setup function, and explicitly set the
first transition to 0. This way the first loop immediately causes the
first transition to happen and everything is more tidy.