Slow down several parameters at once
In the spirit of traceable and atomic commits, this one changes a whole
bunch of parameters all together, all of them randomly and without any
test, to make most effects slower as they felt too fast on the wall.

Also adds the MINUTES macro, which is cute. It makes the code look like
a syntax error.