Fix the flickering bug in PolarSwipe
Finally figured out why it was flickering random colors at the edges:
the outer LEDs were in the band for several frames, and the condition to
select a new color was already triggering while they were in band.

Fixed it by making sure that the new color is changed only when all the
LEDs are out of band, and improved the code a bit while I was at it.