The most insane performance improvement of my life
In the previous commit I added new coordinates structures to each Led so
that they could hold their transformed coordinates, and I noticed a big
drop in performance (~27%).

The only change at runtime was the size in memory of the Led struct,
which had now basically doubled. I realized then that the performance
drop must be due to having to copy ~2x data for each call (structs are
copied, not passed by reference).

So I tried with this commit, where everything is a class now (passed by
reference by default) but most importantly all effects now work directly
with pointers to the Leds.

This removed all the pointless copies and nearly TRIPLED the FPS!