Create the first control hint
This commit creates an enum of control hint flags that the effects and
animations can use to signal info to the control logic.

In this commit only one flag exists, ROTATE_SPACE, that signals to the
control logic that it needs to apply the global rotation transform
before running the animation / effect.

Close #13 (well, not really, but no one can object)