Crude improvements to RGBodyProblem
Now supports 1, 2 or 3 emitters. If 1 or 2, the colors are also now
complementary and not fixed to be RGB (1 emitter still uses a
moodlight).

There's only one reason not to scale it to multiple emitters, and that
is the random complementary colors functions which is now a crude hack
due to the fact that I can't remember how to correctly edit an array.