Add HeatMap to ElectricSparks, give it a chill pill
The heatmap palette looks sick with the sparks effect, so this adds it.
Still thinking about a good way to generate palettes on the fly for it.

Also reduces the frequency of sparks a bit because since I tripled the
framerate it's too fast (did I mention I tripled the framerate?)