Replace LightHouse with NinjaStar
Lighthouse looked rather shitty because of all those sharp edges.
This commit starts with the same objective (multiple rotating beams) and
ends with a much different (and better) effect, where the beams are
smooth because they are a power of a sine wave.

This looks great with many beams because it projects weird shadows on
the wall!