Smooth edge for PolarSwipe
PolarSwipe was looking clunky especially at low speed, because the leds
were either on in the band or off outside of it, resulting in the leds
suddenly popping in clunkily.

This commit makes the brightness linearly dependent on the distance to
the band's center, which gives it a smooth edge and makes it way more
pleasing.