The big migration to PlatformIO
The project is now getting too big and split across too many files
to be manageable in the arduino IDE.

Everyone claims that VSCode is a good editor, and that PlatformIO
is the way to go for arduino projects these days. Let's give it a shot.

It definitely makes it saner to work with so many files, that is for sure.
And it should support the future expansion, leading up to the big ESP32
migration one day.