Add support for Energy control
First step towards #26: creates a static class that holds the global
energy and creates the EnergyParam class that automatically interpolates
the value based on the energy.

Actually using it in effects still creates a lot of issues, and to have
smooth interpolation many will need to use an actual float.
But the infrastructure is here now!