Remove PowerOn from Web, fix bug
It made sense to have both on and off of course, except that Next
already did that in a way as it immediately forced a transition.
Unfortunately it did not reset the ON flag, meaning it would play a
transition and then remain black.

This commit fixes the bug and simplifies the UI in one go, by using the
Next button to also act as On.