Move web server to Core 0
Doesn't really make all that much of a difference in practice (the web
server requests are still handled on Core 1 anyway), but this commit
moves the web server to core 0 to keep it out of the way of the LEDs.

Most likely does not matter at all. The real difference would be to move
the request handlers to Core 0, but I don't understand that yet.

Anyway, in the next commit we'll be sending nearly 4K of HTML in 6ms, so
it makes no difference at all.