Cache static file, revert to single-page
The file system is really nice and cool, but 10x slower than serving a
hardcoded string (understandably).

This commit adds caching, loading the page in memory immediately and
serving it from RAM in just 6ms. It also reverts back to one single HTML
page, which is both faster and easier to manager for my backend-dev ass.