Serve index.html from the file system again
Now that the buttons are using POST messages and prevent reloading,
there is no need to cache the page on the server anymore.

It's slower of course, but more practical for development as I can just
reupload a new file system in theory without even reloading the code.
This also saves about 5kB of RAM, which is not critical but also why
waste it. Serving from RAM takes 5ms, serving from FS takes 50ms, but
with the POST buttons this happens way more infrequently.