Even better-er PerformanceMonitor
Highly optimized frame counter in a ring buffer. The boundaries between
effects and animations will be off, but who cares really (the buffer
should last ~0.64s at 200 fps).

The coolest thing is the FASTLED_SHOW macro, that makes it so we are
literally counting each rendered frame wherever it happens. This allows
us to count the frames even during animations (not possible until now)