First working iteration of AP config
Start in AP mode if there are no stored wifi credentials and show a page
to configure it.

It still has a few issues: to start, I haven't reviewed the code myself
yet (yikes). The configuration page is also considerably different than
the control panel, and if the connection is successful iOS closes
immediately the page and doesn't show the instructions.

But boy LLMs are impressive.